New Components and Updated Logic File
Added a Logout component to authentication that connects itself to the auth logic file and performs the action. This way, you can insert the component where you would like without having to connect it to a logic file.

Changed how the logic file works in Authentication. Instead of letting the reducers doing a lot of the heavy lifting, now sagas will be the ones in charge of performing all the logic needed before an action goes to a reducer. This means a reducer only needs to watch one action, which that action is made to only update that reducer.