File cleanup and new auth component
Removing leftover files to get to a base boilerplate.

Added a component that checks if a user is logged in, if not, redirect to the login page. Also, will check if a localSession exists, and if so, log the user into the store. Finally, if a user becomes logged in, it will redirect the user to the home page.

Added a dumb component called Alert. It wraps the antd version of alert, but has a new option called activate. This will determine if the alert is to show or hide.