Component Updates
Changed checkIfLoggedIn component to do a check of whenever the logged in state is updated. First, there is now an array of strings that serve as a list of 'restricted' paths. Then, if login state is updated to logged in, it will check if the current path is in the list of restricted paths, if so, it will redirect the user back to app home.

Changed SideNav component to automatically selected a menu item if a user lands on a url other then the home url.